Remove redundant casts in GD32 drivers
Clean up redundant casts in GD32 driver code. In lib-gd32/src/gd32_uart.cpp the explicit (uint32_t) casts when writing bytes to USART_TDATA for GD32H7XX were removed (use the masked byte directly). In lib-gd32/src/h/gd32_spi.cpp the static_cast<uint32_t> around SPI_NSS_GPIO_PINx passed to GPIO_BOP/BC was removed. These are cosmetic simplifications that avoid unnecessary casting and potential compiler/warning noise without changing behavior.

+++ lib-gd32/src/h/gd32_spi.cpp
@@ -33,13 +33,13 @@ static uint8_t s_n_chip_select = GD32_SPI_CS0;
 
 static inline void cs_high() {
     if (s_n_chip_select == GD32_SPI_CS0) {
-        GPIO_BOP(SPI_NSS_GPIOx) = static_cast<uint32_t>(SPI_NSS_GPIO_PINx);
+        GPIO_BOP(SPI_NSS_GPIOx) = SPI_NSS_GPIO_PINx;
     }
 }
 
 static inline void cs_low() {
     if (s_n_chip_select == GD32_SPI_CS0) {
-        GPIO_BC(SPI_NSS_GPIOx) = static_cast<uint32_t>(SPI_NSS_GPIO_PINx);
+        GPIO_BC(SPI_NSS_GPIOx) = SPI_NSS_GPIO_PINx;
     }
 }
